Made InteractorWrapper private.
